STLAlloc

+++ Engine/Include/Engine/Core/Interfaces/IMalloc.h
@@ -2,6 +2,6 @@
 
 class Malloc {
 public:
-  virtual void* allocate(size_t bytes, size_t alignment) = 0;
+  virtual void* allocate(uint32_t bytes, uint32_t alignment) = 0;
   virtual void deallocate(void* ptr) noexcept = 0;
 };
